OneDrive/Desktop/Дипломный проект/src/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==0.19.0